requirement2.txt
flask
gtts
SpeechRecognition
langchain
langchain-groq
langchain-pinecone
langchain-ollama
langchain-community
langchain-core
pinecone-client
tqdm
python-dotenv
pypdf